==> analysis/14_Day3/analysis/15_true_and_false/scripts/02_smallest_intron.Rmd <==
---
title: "Smallest intron"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

# Are the significantly retained introns the smallest intron in each gene?

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = F, message = F)
library(tidyverse)
library(kableExtra)
library(ggpubr)
```
 
## sig RI genes VS background genes

These are the sig RI *events* assembled from all three tools. No PSI filtering was added. 

```{r}
# load the sig RI locus in day 3 and check out the length of that RI
# IRfinder : 0 based for start and end
# GTF : 1 based for start and end

# also here I don't have any filtering on PSI or delta PSI
d3RI <- read_csv('../../13_fig1/output/all_tools_sig05_locus_info_day3_with_all_PSI_values.csv') %>%
  filter(GeneID != 'ENSG00000259827') %>%
  dplyr::select(GeneID, locus, type, external_gene_name) %>%
  filter(type == 'RI') %>%
  separate(locus, into = c('chr','coord'), sep = ':', remove = F) %>%
  separate(coord, into = c('start','end'),sep = '-', convert = T) %>%
  mutate(RIlen = end - start)


# collapse the df to see the genes that have more than one RI event
sig <- d3RI %>%
  group_by(GeneID) %>%
  arrange(RIlen, .by_group = TRUE) %>%
  summarise(RI_len_all = paste0(RIlen, collapse = ","),
            locus_all = paste0(locus, collapse = ","),
            cnt = n()) %>%
  arrange(desc(cnt)) %>%
  rename(ensembl_gene_id = GeneID) %>%
  mutate(ensembl_gene_id = str_replace_all(string = ensembl_gene_id, pattern = '\\..*$', replacement = ''))
```

```{r}
# load the data for the background introns
# the genes are the ones that have expression in all 9 samples
# filter out ones that have intronLen <= 30, since they are deduced from sequences, could simply be indels 
bg <- read_csv('../data/unique_introns_for_expressive_genes.csv') %>%
  filter(intronLen > 30)
```

```{r}
# select the shortest intron in each gene
bg_short <- bg %>%
  group_by(ensembl_gene_id) %>%
  summarise( shortest = min(intronLen)) 

# join that with the sig table
res  <- sig %>%
  dplyr::select(ensembl_gene_id, RI_len_all, locus_all) %>%
  left_join(bg_short , by = 'ensembl_gene_id') %>%
  mutate(min_sig_len = str_replace_all(string = RI_len_all, pattern = '\\,.*$', replacement = ''), 
         min_sig_len= as.numeric(min_sig_len) )  %>%
  mutate(abs_len_diff = abs(min_sig_len - shortest))


```


```{r}
res %>%
  head() %>%
  kable(caption = 'A look at the joined table') %>%
  kable_styling()
```

Are all those genes exist in the background ones?

```{r}
abundant <- read_csv('../../11_salmon_quant/data/raw_salmon_cts_day3_matrix_nonzero_all_gt0.csv') %>%
  mutate(Name = str_replace_all(string = Name, pattern = '\\..*$', replacement = '')) 

library(VennDiagram)

diff_list <- list(sig_RI_gene = res$ensembl_gene_id,
                  background = abundant$Name )
fill <- c("light blue", "pink")
size  <- rep(0.5, 2)
venn <- venn.diagram(x = diff_list, 
      filename = NULL,
      height = 2000,
      width = 2000, fill = fill,
      cat.default.pos = "text", 
      cat.cex = size,
      main = "Are all the sig genes in the background ones");
png('../figs/sig in bg.png', width = 4, height = 4, units = 'in', res = 300)
grid.draw(venn)
dev.off()
```
```{r out.width=400}
knitr::include_graphics('../figs/sig in bg.png')
```

Yes, all the *genes* that have any sig RI events exist in the background ones. Again, we are keeping all the genes that have at least one non-zero count in the count matrix, meaning any gene that shows any indication of expression will be in the background ones. 

## Smallest intron in that gene VS smallest sig RI event in that gene

The major metric here is the *abs_len_diff*, which is absolute value of (smallest intron length in that gene  minus  smallest significantly retained intron length in that gene).


```{r}
# show the distribution of the abs_len_diff
res %>%
  gghistogram(x = 'abs_len_diff', bins = 200,
              color = 'red',
              xlab = 'absolute length difference',
              title = 'Histogram of absolute length difference\nbetween the shortest intron in the gene and the\nshortest retained intron identified by IRfinder') +
  ggsave('../figs/Histogram of absolute length difference.jpg', width = 8, height = 6)
```


### What about some sig RI event that are very long retained introns.

A look in the IGV (loading the annotation file used in running those tools). The lines that have a pink line below it are the long introns that are retained. 

```{r out.width=600}
# what is happening at the abs_len_diff > 6000?
# add screenshots form IGV. they look real from the annotation file
file_path <- list.files('../figs/IGV_screenshots _for_long_sig_RI/', full.names = T)
knitr::include_graphics(file_path)
```



```{r}
### Zooming in 
# a zoom in to look at only the abs_len_diff < 500
p <- res %>%
  filter(!is.na(shortest)) %>%
  filter(abs_len_diff < 500) %>%
  gghistogram(x = 'abs_len_diff', bins = 200,
              color = 'red',
              xlab = 'absolute length difference',
              title = 'Same plot but absolute length difference less than 500') +
  ggsave('../figs/Distribution of absolute length difference less than 500.jpg', width = 8, height = 6)


# a table look at the interval of width 20
interval20 <- res %>%
  filter(!is.na(shortest)) %>%
  filter(abs_len_diff < 500) %>%
  pull(abs_len_diff)  %>%
  cut_interval(., length=20) %>%
  table %>%
  as_tibble %>%
  rename(interval = names(.)[1],
         cnt = names(.)[2])
```

### Tallying of the abs_len_diff

```{r}


# group by the abs_len_diff and then do a bar plot (abs_len_diff < 20)
res %>%
  filter(!is.na(shortest)) %>%
  group_by(abs_len_diff) %>%
  summarise(cnt = n()) %>%
  filter(abs_len_diff < 20) %>%
  ggbarplot(x = 'abs_len_diff', y = 'cnt',
            color = 'red', fill = 'red', label = T,
            title = 'Tally of absolute len difference (< 20)',
            xlab = 'absolute length difference')  +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        axis.text=element_text(size=10)) +
  ggsave('../figs/Tally of absolute len difference when it is below 20.jpg', width = 6, height = 5)

# group by the abs_len_diff and then do a bar plot (all the different abs_len_diff)
res %>%
  filter(!is.na(shortest)) %>%
  group_by(abs_len_diff) %>%
  summarise(cnt = n()) %>%
  ggbarplot(x = 'abs_len_diff', y = 'cnt',
            color = 'red',fill = 'red',
            title = 'Tally of absolute len difference',
            xlab = 'absolute length difference')  +
  theme(axis.text.x = element_text(angle = 45, hjust = 1),
        axis.text=element_text(size=5)) +
  ggsave('../figs/Tally of absolute len difference.jpg', width = 36, height = 5)

# it's too ridiculously wide, thinking of how to split that to 2 equal parts 
tally_table <- res %>%
  filter(!is.na(shortest)) %>%
  group_by(abs_len_diff) %>%
  summarise(cnt = n()) %>%
  mutate(Csum = cumsum(cnt),
         half_sum = floor(sum(.$cnt)/2),
         half = if_else(Csum < half_sum, 'first','second'))


# draw again with the wide bars split into 2
tally_table %>%
  ggbarplot(x = 'abs_len_diff', y = 'cnt',
            color = 'red',label = T,
            title = 'Tally of absolute len difference',
            xlab = 'absolute length difference')  +
  facet_grid(half ~ ., scales="free_x") +
  theme(axis.text.x = element_text(angle = 90, hjust = 1),
        axis.text=element_text(size=5)) +
  ggsave('../figs/Tally of absolute len difference split.jpg', width = 18, height = 5)



# have a peek of the tally_table
tally_table %>%
  dplyr::select(abs_len_diff, cnt) %>%
  head(n = 10) %>%
  kable() %>%
  kable_styling(full_width = F)
```

Of all the genes that have sig RI events, what percentage of them has a RI event that is the smallest intron in that gene? (One gene may have multiple RI events. If there exist one event that is the smallest intron in the gene, it will be counted. )

```{r}


res %>%
  filter(!is.na(shortest)) %>%
  group_by(abs_len_diff) %>%
  summarise(cnt = n()) %>% 
  mutate(thre20 = abs_len_diff < 20) %>% 
  split(.$thre20) %>%
  map_dfc( ~ sum(.$cnt)) %>%
  rename(gt20 = names(.)[1],
         lt20 = names(.)[2]) %>%
  kable(caption = 'How many abs_len_diff < 20 and how many are > 20') %>%
  kable_styling( position = 'left')



res %>%
  filter(abs_len_diff == 0) %>%
  summarise(Num = nrow(.)) %>%
  mutate(Total = nrow(res)) %>%
  mutate(Percentage = paste(round(Num/Total*100, 2), '%', sep = '')) %>%
  kable(caption = 'what percentage of geneshas a RI event that is the smallest intron in that gene?') %>%
  kable_styling(full_width = F, position = 'left')
  
  
```

 


```{r}
# write out the res table
res %>%
  write_csv('../data/sig RI genes VS background.csv')
```
