==> analysis/14_Day3/analysis/15_true_and_false/scripts/08_the_bimodal_ss5_score.Rmd <==
---
title: "The seq logo for the MaxEnt splice site"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(tidyverse)
library(kableExtra)
library(ggpubr)
library(ggseqlogo) 
```

For day 3.

```{r}
# load the day 3 SS score files (intersect sig)
data_path <- '../../05_IRfinder/output/'
fns <- list.files(data_path, 'txt', full.names = T)

ss5 <- read_delim(fns[1],comment = '>',delim = '\t',col_names = F) %>%
  dplyr::select(seq = X1,
                score = X2) %>%
  mutate(score = str_replace(score, 'MAXENT: ',''),
         score = as.numeric(score)) 
  

nrow(ss5)

ss5 %>%
  gghistogram(x = 'score')


res <- read_csv('../../05_IRfinder/data/02_intersect_sig_and_intersect_non_sig_all_other_info.csv') 


res %>%
  ggplot() + 
  geom_logo(ss5_seq ) + 
  theme_logo() 


res %>%
  split(.$Significance) %>%
  map(~ ggplot() + 
        geom_logo( .$ss5_seq ) + 
        theme_logo() + 
        labs(title = str_glue('{.$Significance} logo')))
```

```{r}
# do a seq logo of the top and bottom 1000
N <- 50

ss5_both_ends <- bind_rows(
  ss5 %>%
    arrange(desc(score)) %>%
    top_n(N) %>%
    mutate(group = 'top'),
  ss5 %>%
    arrange(desc(score)) %>%
    top_n(-N) %>%
    mutate(group = 'bottom')
)
```

```{r}
# do a seq logo
ss5_both_ends %>%
  split(.$group) %>%
  map(~ ggplot() + 
        geom_logo( .$seq ) + 
        theme_logo() + 
        labs(title = str_glue('{.$group} logo'))+
        theme_classic() +
        theme(
              text = element_text(size=25),
              axis.text.x = element_text( hjust = 1, size = 15))+
        ggsave(str_glue('../figs/ss3 intersect {.$group} logo.pdf'),device = 'pdf', width = 18, height = 10, dpi = 300))

```
