==> analysis/14_Day3/analysis/15_true_and_false/scripts/09_intron_rank_histogram.Rmd <==
---
title: "Histogram of intron rank"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,message = F, warning = F)
library(tidyverse)
library(kableExtra)
library(ggpubr)
```

Steps:

+ Get all of the intron in a gene
+ Get the largest RI in that gene
+ Get the rank of the 2 in 1


```{r}
# get the sig events (the intersection)
sig <- read_csv('../../05_IRfinder/data/01_intersect_sig_and_intersect_non_sig_locus.csv') %>%
  filter(Significance == 'Sig')

sig_gene <- sig %>%
  distinct(GeneID) %>%
  pull(GeneID)


# get the largest sig RI len in that gene
largest_RI <- sig %>%
  mutate(len = end - start) %>%
  group_by(GeneID) %>%
  summarise(largest = max(len))
```


```{r}
# get all of the introns of the genes that have sig RI events, and sort within each gene the intronLen in an ascending order
bg <- read_csv('../data/unique_introns_for_expressive_genes.csv') %>%
  filter(intronLen > 30) %>%
  filter(ensembl_gene_id %in% sig_gene) %>%
  group_by(ensembl_gene_id) %>%
  arrange(intronLen, .by_group =T)

# join them together first

ret <- bg %>%
  split(.$ensembl_gene_id) %>%
  map_df(~ mutate(.data = ., rankIntron = rank(intronLen, ties.method = 'first') )) %>%
  left_join(largest_RI %>%
              dplyr::rename(ensembl_gene_id = GeneID), by = 'ensembl_gene_id')

# it's suggested to divide the raw rank by the total number of introns in that gene, so that we know if it's the smallest intron in many introns or the only intron.
total <- ret %>%
  group_by(ensembl_gene_id) %>%
  summarise(total = max(rankIntron))

# get the rank of the largest RI event
rank_RI <- ret %>%
  filter(intronLen == largest) %>%
  distinct(ensembl_gene_id, .keep_all = T)

# join with the total number of intron table
final <- rank_RI %>%
  full_join(total, by = 'ensembl_gene_id') %>%
  mutate(ratio = round(rankIntron/total, 2))
  
```

```{r}
# plot the histogram of the ratio
bin_num <- 20

final %>%
  gghistogram(x = 'ratio', 
              bins = bin_num, fill = '#00468B', color = 'white',
              xlab = 'Ratio: rank of the longest retained intron/total num of intron in that gene',
              title = str_glue('Histogram of the ratio bin number {bin_num}')) +
  ggsave(str_glue('../figs/Histogram of the ratio bin number {bin_num}.jpg'), width = 8, height = 5, dpi = 300)
```

```{r}
# a bar plot
rank_RI %>%
  group_by(rankIntron) %>%
  summarise(cnt = n()) %>%
  ggbarplot(x = 'rankIntron', y = 'cnt', xlab = 'rank of the longest retained intron',
            fill = '#00468B', color = 'white',
            title = 'Bar plot of the rank of the longest retained intron') +
  ggsave('../figs/bar plot of the rank of the longest retained intron.jpg', width = 10, height = 6, dpi = 300)
```
