==> analysis/14_Day3/analysis/15_true_and_false/scripts/10_scrutize_the_SS_score.Rmd <==
---
title: "Scrutinize the SS score"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F, warning = F)
library(ggpubr)
library(kableExtra)
library(tidyverse)
library(GenomicRanges)
library(BSgenome)
library(BSgenome.Hsapiens.NCBI.GRCh38)
library(seqinr)
```

```{r}
# load all of the bg introns seq info
all_bg <- read_csv('../../05_IRfinder/data/03_non_AS_seq_info.csv')
```

## SS5 

(The 4,5 should be G,T)

```{r}
# make it a function of retrieving the ss5 seq from the score that is the input
genome <- BSgenome.Hsapiens.NCBI.GRCh38
seqlevelsStyle(genome) = "UCSC"

get_ss5_seq_of_score <- function(Score) {
  upper = Score + 0.2
  lower = Score - 0.2
  
  df = all_bg %>% 
    filter(ss5_score <= upper & ss5_score >= lower)
  
  # retrieve the ss5 sequences
  grange5 <-  makeGRangesFromDataFrame(df, 
                                       seqnames.field=c("chr"),  
                                       start.field="ss5_start", 
                                       end.field=c("ss5_end"), 
                                       strand.field="strand")
  df <- df %>% 
    mutate(ss5_seq = as.character(BSgenome::getSeq(genome, grange5))) %>% 
    dplyr::select(ss5_score, ss5_seq, ss3_score, category)
  
  return(df)
}


# the percentage of the seqs that have GT at 4,5
get_correct_seqs_perc <- function(Score){
  ret = get_ss5_seq_of_score(Score)
  perc = ret %>% 
    pull(ss5_seq) %>% 
    str_locate('GT') %>% 
    as.data.frame %>% 
    filter(start == 4 & end == 5) %>% 
    summarise(correct_seqs_perc = round(nrow(.)/nrow(ret)*100,2)) %>% 
    pull(correct_seqs_perc)
  return(perc)
}
```

```{r}
# the lowest score in ss5_score
min_ss5_score <- min(all_bg$ss5_score)
   
# when score == 0
score0 <- 0

# when score is half of min_ss5_score
score1 <- min_ss5_score/2

# the next search will be the middle point between min_ss5_score and score1
score2 <- (min_ss5_score + score1)/2

# the next search will be the middle point between score2 and score1
score3 <- (score2 + score1)/2

# would like to see the middle point between score1 and 0 too
score4 <- score1/2

# would like to see the middle point between score4 and 0
score5 <- score4/2

# the middle point between score4 and score5
score6 <- (score4 + score5)/2

# wait a min, why am I writing so much repetative code?
scores <- c(min_ss5_score, score0, score1, score2, score3, score4, score5, score6)

res_ss5 <- scores %>% 
  set_names(scores) %>% 
  map_dfc(~ get_correct_seqs_perc(.)) %>% 
  gather(key = Score, value = Percentage) %>% 
  mutate(Score = as.numeric(Score)) %>% 
  arrange(Score)
```

```{r}
res_ss5 %>% 
  kable(caption = 'Percentage of seqs with correct SS5
        signal at the score threshold') %>% 
  kable_styling(full_width = F)
```


```{r}
# so the plan is to remove any of the seqs that have a score below the threshold where half of the seqs don't have the correct signal at that place
guess <- seq(-8, -6, by = 0.2)
guess_res <- guess %>% 
  set_names(guess) %>% 
  map_dfc(~ get_correct_seqs_perc(.)) %>% 
  gather(key = Score, value = Percentage) %>% 
  mutate(Score = as.numeric(Score)) %>% 
  arrange(Score)

```

-7.2 score corresponds to the percentage being 50%. So that means seqs that have score in range [-7.4, -7.0] have 50% of them have the correct signal at the location.


## SS3


```{r}
# slightly change the function to accomodate the ss3 part
# make it a function of retrieving the ss5 seq from the score that is the input

get_ss3_seq_of_score <- function(Score) {
  upper = Score + 0.2
  lower = Score - 0.2
  
  df = all_bg %>% 
    filter(ss3_score <= upper & ss3_score >= lower)
  
  # retrieve the ss3 sequences
  grange3 <-  makeGRangesFromDataFrame(df, 
                                       seqnames.field=c("chr"),  
                                       start.field="ss3_start", 
                                       end.field=c("ss3_end"), 
                                       strand.field="strand")
  df <- df %>% 
    mutate(ss3_seq = as.character(BSgenome::getSeq(genome, grange3))) %>% 
    dplyr::select(ss3_seq, ss3_score, category)
  
  return(df)
}


# the percentage of the seqs that have AG at 19,20
get_correct_seqs_perc_ss3 <- function(Score){
  ret = get_ss3_seq_of_score(Score)
  perc = ret %>% 
    pull(ss3_seq) %>% 
    str_locate('AG') %>% 
    as.data.frame %>% 
    filter(start == 19 & end == 20) %>% 
    summarise(correct_seqs_perc = round(nrow(.)/nrow(ret)*100,2)) %>% 
    pull(correct_seqs_perc)
  return(perc)
}
```

```{r}
# the lowest score in ss3_score
min_ss3_score <- min(all_bg$ss3_score)
   
# when score == 0
score0 <- 0

# when score is half of min_ss3_score
score1 <- min_ss3_score/2

# the next search will be the middle point between min_ss3_score and score1
score2 <- (min_ss3_score + score1)/2

# the next search will be the middle point between score2 and score1
score3 <- (score2 + score1)/2

# would like to see the middle point between score1 and 0 too
score4 <- score1/2

# would like to see the middle point between score4 and 0
score5 <- score4/2

# the middle point between score4 and score5
score6 <- (score4 + score5)/2

# wait a min, why am I writing so much repetative code?
scores <- c(min_ss3_score, score0, score1, score2, score3, score4, score5, score6)

res_ss3 <- scores %>% 
  set_names(scores) %>% 
  map_dfc(~ get_correct_seqs_perc_ss3(.)) %>% 
  gather(key = Score, value = Percentage) %>% 
  mutate(Score = as.numeric(Score)) %>% 
  arrange(Score)
```

```{r}
res_ss3 %>% 
  kable(caption = 'Percentage of seqs with correct SS3 signal at the score threshold') %>% 
  kable_styling(full_width = F)
```

```{r}
# same to do the process at SS3
guess <- seq(1.1, 1.4, by = 0.05)
guess_res_ss3 <- guess %>% 
  set_names(guess) %>% 
  map_dfc(~ get_correct_seqs_perc_ss3(.)) %>% 
  gather(key = Score, value = Percentage) %>% 
  mutate(Score = as.numeric(Score)) %>% 
  arrange(Score)
```

```{r}
# get the score threshold intended programatically
two <- list(ss5 = guess_res, ss3 = guess_res_ss3)
two %>% 
  map_dfc(~ mutate(.data = ., diff = abs(50 - Percentage)) %>%
        filter(diff == min(.$diff)) %>% 
        pull(Score))
```
